CAMEL-10389: Moved string related function from ObjectHelper to StringHelper
